added content for worship and missions
funny thing, the jumbotron is pushed to the bottom for some reason, it might be a php error

also we need to get images for these sections, adding some color will not fix the bland. Similiar to what I did with the discipleship page

and btws the disc page works fine. the content is in the php content tags. you told me to ignore that and put my html after the PHP end tags.

so dont, give me a told you so tomorrow....

also its 3:30 am, yes... AM

have a great day/evening

I'm gonna go sleep..............ZZZZZZzzzzzzz.....